+++ spec/internal/app/models/schema_migration.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+class SchemaMigration < ActiveRecord::Base
+end