+++ spec/internal/app/presenters/application_presenter.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+class ApplicationPresenter < Magic::Presenter::Base
+	# Define presentation-specific methods here.
+	# They will be shared by all the descendant presenters.
+end