+++ spec/internal/app/presenters/person_presenter.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class PersonPresenter < ApplicationPresenter
+	def name = "#{first_name} #{last_name}"
+end