+++ spec/internal/db/schema.rb
@@ -11,3 +11,22 @@
 # migrations use external dependencies or application code.
 #
 # It's strongly recommended that you check this file into your version control system.
+
+ActiveRecord::Schema[7.2].define(version: 2024_10_26_111752) do
+	create_table 'people', force: :cascade do |t|
+		t.integer 'parent_id'
+		t.string 'first_name'
+		t.string 'last_name'
+		t.datetime 'created_at', null: false
+		t.datetime 'updated_at', null: false
+		t.index ['parent_id'], name: 'index_people_on_parent_id'
+	end
+
+	create_table 'users', force: :cascade do |t|
+		t.string 'name'
+		t.datetime 'created_at', null: false
+		t.datetime 'updated_at', null: false
+	end
+
+	add_foreign_key 'people', 'people', column: 'parent_id'
+end